Reorganize container components.
Don't track containers in their own submodule or elide part of the
component name in the path. Instead, make module file names match the
names of their components, and track containers alongside their
presentational components.
